updated tpwc ai sup. to release 3.01

+++ tpwcas/tpwcas_debug.sqf
@@ -14,19 +14,19 @@ tpwcas_fnc_debug =
 	while { true } do
 		{
 			{
+			if ( local _x ) then 
+				{
 				if( isNil { _x getVariable "tpwcas_debug_ball" } ) then
 					{
 					_ball = createVehicle ["Sign_sphere25cm_EP1", [(random 15),(random 15),1], [], 0, "NONE"];
 					_ball setObjectTexture [0,"#(argb,8,8,3)color(0.99,0.99,0.99,0.7,ca)"];  // white
-					
-					if ( isDedicated ) then 
-					{
-						_nul = ["suppressDebug", [_x, _ball, 0]] call CBA_fnc_globalEvent;
-					};
-					_colur = "ColorWhite";					
-					if( (side _x) getFriend WEST < 0.6 ) then 
+					if ( tpwcas_multi_player ) then 
 						{
-						_color = "ColorRed"; 
+							_nul = ["suppressDebug", [_x, _ball, 0]] call CBA_fnc_globalEvent;
+						};
+					_color = "ColorWhite";					
+					if( (side _x) getFriend WEST < 0.6 ) then 
+						{ _color = "ColorRed"; 
 						} 
 					else 
 						{
@@ -59,13 +59,13 @@ tpwcas_fnc_debug =
 					_marker setMarkerType "mil_triangle";
 					_marker setMarkerDir (getDir _x);
 					};
-			
-			if( !( isNull _x ) && alive _x ) then // better to double check for unit being alive ...
-				{
-				_ball = _x getVariable "tpwcas_debug_ball";
-				_level = _x  getVariable ["tpwcas_supstate", 0];
-				_ball_level = _x getVariable ["tpwcas_ball_state", 0];
-			
+				
+				if( !( isNull _x ) && alive _x ) then // better to double check for unit being alive ...
+					{
+					_ball = _x getVariable "tpwcas_debug_ball";
+					_level = _x  getVariable ["tpwcas_supstate", 0];
+					_ball_level = _x getVariable ["tpwcas_ball_state", 0];
+				
 				if !( _level == _ball_level ) then 
 					{
 					switch ( true ) do
@@ -73,7 +73,7 @@ tpwcas_fnc_debug =
 						case ( fleeing _x ): 
 							{  
 							_ball setObjectTexture [0,"#(argb,8,8,3)color(0.0,0.0,0.0,0.9,ca)"];  // black
-							if ( isDedicated ) then 
+							if ( tpwcas_multi_player ) then 
 								{
 								["suppressDebug", [_x, _ball, 1]] call CBA_fnc_globalEvent;
 								};
@@ -90,7 +90,7 @@ tpwcas_fnc_debug =
 						case ( _level == 1 ): 
 							{  
 							_ball setObjectTexture [0,"#(argb,8,8,3)color(0.1,0.9,0.1,0.7,ca)"];  // green
-							if ( isDedicated ) then 
+							if ( tpwcas_multi_player ) then 
 								{
 								_msg = format["'suppressDebug' trigger sent for unit [%1] - value [%2] - ball [%3]", _x, 2, _ball];
 								[ _msg, 9 ] call bdetect_fnc_debug;
@@ -103,7 +103,7 @@ tpwcas_fnc_debug =
 						case ( _level == 2): 
 							{  
 							_ball setObjectTexture [0,"#(argb,8,8,3)color(0.9,0.9,0.1,0.7,ca)"]; //yellow
-							if ( isDedicated ) then 
+							if ( tpwcas_multi_player ) then 
 								{
 								_nul = ["suppressDebug", [_x, _ball, 3]] call CBA_fnc_globalEvent;
 								};
@@ -114,7 +114,7 @@ tpwcas_fnc_debug =
 						case ( _level == 3 ): 
 							{  
 							_ball setObjectTexture [0,"#(argb,8,8,3)color(0.9,0.1,0.1,0.7,ca)"]; //red  
-							if ( isDedicated ) then 
+							if ( tpwcas_multi_player ) then 
 								{
 								_nul = ["suppressDebug", [_x, _ball, 4]] call CBA_fnc_globalEvent;
 								};
@@ -127,18 +127,18 @@ tpwcas_fnc_debug =
 						
 					};
 				};
-								
-			if ( lifestate _x != "ALIVE" ) then
-				{ 
-				_ball = _x getVariable "tpwcas_debug_ball";				
-				detach _ball;
-				_ball setPosATL [(random 15),(random 15),0];
-				};
-				
+									
+				if ( lifestate _x != "ALIVE" ) then
+					{ 
+					_ball = _x getVariable "tpwcas_debug_ball";				
+					detach _ball;
+					_ball setPosATL [(random 15),(random 15),0];
+					};
+				};	
 			} foreach allunits;
 			
 			{
-				if( ! ( isNil { _x getVariable "tpwcas_debug_ball" } ) ) then
+				if( (local _x) && !( isNil { _x getVariable "tpwcas_debug_ball" } ) ) then
 				{
 					deleteVehicle ( _x getVariable "tpwcas_debug_ball" );
 					deleteMarker( _x getVariable "tpwcas_debug_marker" );
